Use less precise divisor for age calculation
Using 365.25 as a divisor seems to result in age calculations (by years)
that are off by a year the older someone is.  I am unsure if this
problem lies in the datetime library, or if the math is right but the
outcome differs from what we as humans define as "X years old".

Given more time, I would look more into this.